Refactor out functions to classes that make sense
