--- backend/molgenis-emx2-semantics/src/main/java/org/molgenis/emx2/semantics/graphgenome/HTTPGet.java
@@ -1,36 +1,0 @@
-package org.molgenis.emx2.semantics.graphgenome;
-
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-
-public class HTTPGet {
-
-  /**
-   * Simple function to perform an HTTP GET request on a URL.
-   *
-   * @param url
-   * @return
-   * @throws Exception
-   */
-  public static String httpGet(String url) throws Exception {
-    URL obj = new URL(url);
-    HttpURLConnection con = (HttpURLConnection) obj.openConnection();
-    con.setRequestMethod("GET");
-    con.setRequestProperty("User-Agent", "Mozilla/5.0");
-    int responseCode = con.getResponseCode();
-    if (responseCode == HttpURLConnection.HTTP_OK) {
-      BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
-      String inputLine;
-      StringBuffer response = new StringBuffer();
-      while ((inputLine = in.readLine()) != null) {
-        response.append(inputLine);
-      }
-      in.close();
-      return response.toString();
-    } else {
-      throw new Exception("GET request failed on " + url);
-    }
-  }
-}
